+++ src/app/locales/en/components/share-show.js
@@ -11,6 +11,8 @@ import markdownEditor from './share-show/markdown-editor';
 import urlInfoHandle from './share-show/url-info-handle';
 import urlInfoShare from './share-show/url-info-share';
 import urlInfoRest from './share-show/url-info-rest';
+import edmEditor from './share-show/edm-editor';
+import metadataEditorFooter from './share-show/metadata-editor-footer';
 
 export default {
   tabs: {
@@ -37,4 +39,6 @@ export default {
   urlInfoHandle,
   urlInfoShare,
   urlInfoRest,
+  edmEditor,
+  metadataEditorFooter,
 };