+++ src/app/locales/en/components/share-show/metadata-editor-footer.js
@@ -1,0 +1,6 @@
+export default {
+  publishInfo: 'The dataset will be made available to the public using <strong>{{handleServiceName}}</strong> handle service (ID: <code>{{handleServiceId}}</code>). The unique identifier and the public link will be automatically assigned. <strong>Please keep in mind that once published, your dataset should not be removed.</strong>',
+  cancel: 'Cancel',
+  publish: 'Publish',
+  asOpenData: 'as Open Data',
+};