+++ src/app/locales/en/components/share-show/open-data-editor.js
@@ -1,9 +1,7 @@
 // NOTE: This is a mixin to extend open-data-common translations
 export default {
-  publishInfo: 'The dataset will be made available to the public using <strong>{{handleServiceName}}</strong> handle service (ID: <code>{{handleServiceId}}</code>). The unique identifier and the public link will be automatically assigned. <strong>Please keep in mind that once published, your dataset should not be removed.</strong>',
   publishingData: 'publishing as Open Data',
   publish: 'Publish',
-  asOpenData: 'as Open Data',
   xmlTextareaPlaceholder: 'Type Dublin Core XML metadata here...',
   add: 'Add another',
   inputPlaceholder: 'Enter {{type}}...',