+++ src/app/locales/en/components/share-show/open-data-preview.js
@@ -1,3 +1,3 @@
 export default {
-  noMetadata: 'This dataset has no Dublin Core metadata attached.',
+  noMetadata: 'This dataset has no metadata attached.',
 };